Drop build cache and locking.
Besides being simpler, the build cache is a bad idea in Lisp because
of the presence of restartable errors. If all errors terminated the
build, there would be no problem; but since you can change things and
continue from an error, the cache is not valid.

The current implementation of parallelism should, by this point, be
considered a failed experiment.